requirements.txt
faiss-cpu>=1.7.4
selenium>=4.11.2
webdriver-manager>=4.0.0
crewai>=0.30.0
langchain>=0.1.0
langchain-openai>=0.0.2
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
pandas>=2.0.0
Flask>=2.0.0
Werkzeug>=2.0.0
HEAD
gunicorn>=20.1.0
7a0c1a72 (Initial commit)
